requirements.txt
apache-airflow==3.1.5
psycopg2-binary
requests
pandas
SQLAlchemy
